# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: School Report Card Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /terms:
    get:
      operationId: listTerms
      tags: [terms]
      summary: List terms
      parameters:
        - name: academicYearId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of terms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Term"
    post:
      operationId: createTerm
      tags: [terms]
      summary: Create term
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TermInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Term"
  /terms/{id}:
    get:
      operationId: getTerm
      tags: [terms]
      summary: Get term
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Term
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Term"
    patch:
      operationId: updateTerm
      tags: [terms]
      summary: Update term
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TermUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Term"
    delete:
      operationId: deleteTerm
      tags: [terms]
      summary: Delete term
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  # ── CLASSES ──────────────────────────────────────────────────────────────
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (admin only)
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [admin, teacher, parent]
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  # ── TEACHERS ─────────────────────────────────────────────────────────────
...
  /teachers/{id}:
    get:
      operationId: getTeacher
      tags: [teachers]
      summary: Get teacher
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Teacher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"

  # ── TEACHER ASSIGNMENTS ───────────────────────────────────────────────────
  /teacher-assignments:
    get:
      operationId: listTeacherAssignments
      tags: [teachers]
      summary: List teacher assignments
      parameters:
        - name: teacherId
          in: query
          schema:
            type: integer
        - name: termId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of teacher assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeacherAssignment"
    post:
      operationId: createTeacherAssignment
      tags: [teachers]
      summary: Assign teacher to class subject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeacherAssignmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherAssignment"
  /teacher-assignments/{id}:
    delete:
      operationId: deleteTeacherAssignment
      tags: [teachers]
      summary: Remove teacher assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  # ── STUDENTS ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
...
    Term:
      type: object
      required: [id, academicYearId, name, isCurrent]
      properties:
        id:
          type: integer
        academicYearId:
          type: integer
        academicYearLabel:
          type: string
        name:
          type: string
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        isCurrent:
          type: boolean
    TermInput:
      type: object
      required: [academicYearId, name]
      properties:
        academicYearId:
          type: integer
        name:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        isCurrent:
          type: boolean
    TermUpdate:
      type: object
      properties:
        name:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        isCurrent:
          type: boolean

    # ── Class ─────────────────────────────────────────────────────────────
...
    User:
      type: object
      required: [id, fullName, email, role, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, teacher, parent]
        createdAt:
          type: string
    UserInput:
      type: object
      required: [fullName, email, password, role]
      properties:
        fullName:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        role:
          type: string
          enum: [admin, teacher, parent]
        staffId:
          type: string
        phone:
          type: string
    UserUpdate:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, teacher, parent]

    # ── Teacher ───────────────────────────────────────────────────────────
    Teacher:
      type: object
      required: [id, userId, fullName, email]
      properties:
        id:
          type: integer
        userId:
          type: integer
        fullName:
          type: string
        email:
          type: string
        staffId:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
    TeacherAssignment:
      type: object
      required: [id, teacherId, classSubjectId, termId]
      properties:
        id:
          type: integer
        teacherId:
          type: integer
        teacherName:
          type: string
        classSubjectId:
          type: integer
        className:
          type: string
        subjectName:
          type: string
        termId:
          type: integer
        termName:
          type: string
    TeacherAssignmentInput:
      type: object
      required: [teacherId, classSubjectId, termId]
      properties:
        teacherId:
          type: integer
        classSubjectId:
          type: integer
        termId:
          type: integer

    # ── Student ───────────────────────────────────────────────────────────
...